tests: create a global `conftest.py` (Bug 1942851)
Create a global `conftest.py` for fixtures to be used across
apps. Move test fixtures that will be used in the headless API
into the global conftest.
